Workrounds for Chrome and Firefox printing bugs in documentation and examples
There appears to be a bug in Chrome where is a page element is the same
size as the physical page (from @page) then it actually overflows the
page. Firefox has a long-standing bug whereby absolutely positioned
elements on second and subsequent pages don't render properly when
printed: a non-zero top margin seems to work around this.